trying a lorenz attractor demo; try hot update
