+++ shaders/compute-lorenz.wgsl
@@ -1,0 +1,50 @@
+struct Particle {
+  pos: vec3<f32>,
+  vel: vec3<f32>,
+}
+
+struct SimParams {
+  delta_t: f32,
+  rule1_distance: f32,
+  rule2_distance: f32,
+  rule3_distance: f32,
+  rule1_scale: f32,
+  rule2_scale: f32,
+  rule3_scale: f32,
+}
+
+struct Particles {
+  particles: array<Particle>,
+}
+
+@binding(0) @group(0) var<uniform> params: SimParams;
+@binding(1) @group(0) var<storage, read> particles_a: Particles;
+@binding(2) @group(0) var<storage, read_write> particles_b: Particles;
+
+const tau = 10f;
+const rou = 28f;
+
+fn lozrenz(p: vec3f, dt: f32) -> vec3f {
+  let beta = 8.0 / 3.0;
+  let dx = tau * (p.y - p.x);
+  let dy = p.x * (rou - p.z) - p.y;
+  let dz = p.x * p.y - beta * p.z;
+  return p + vec3<f32>(dx, dy, dz) * dt;
+}
+
+// https://github.com/austinEng/Project6-Vulkan-Flocking/blob/master/data/shaders/computeparticles/particle.comp
+@compute @workgroup_size(64)
+fn main(@builtin(global_invocation_id) GlobalInvocationID: vec3<u32>) {
+  var index = GlobalInvocationID.x;
+
+  var v_pos = particles_a.particles[index].pos;
+  var v_vel = particles_a.particles[index].vel;
+
+  let pos_next = lozrenz(v_pos, params.delta_t * 0.3);
+
+
+  // Write back
+  particles_b.particles[index].pos = pos_next;
+  // particles_b.particles[index].vel = v_vel;
+
+}
