+++ src/global.d.ts
@@ -1,3 +1,5 @@
+/// <reference types="vite/client" />
+
 declare module "*.wgsl" {
   const value: string;
   export default value;
