+++ src/globals.mts
@@ -1,5 +1,7 @@
 import { Atom } from "@triadica/touch-control";
 
+console.log("loading globals.mts");
+
 export var atomDevice: Atom<GPUDevice> = new Atom(null);
 export var atomContext: Atom<GPUCanvasContext> = new Atom(null);
